requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
beautifulsoup4==4.11.1
Django==4.2.7
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap4==23.2
django-crispy-forms==1.14.0
django-livereload-server==0.4
mysqlclient==2.1.1
Pillow==9.2.0
psycopg2==2.9.3
pydantic==2.5.2
pydantic_core==2.14.5
python-environ==0.4.54
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tornado==6.2
typing_extensions==4.8.0
tzdata==2022.1

pip~=22.0.4
attrs~=23.2.0
cryptography~=41.0.7
channels~=4.0.0
djangorestframework~=3.14.0